# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AutWit Copilot API
  version: 0.1.0
  description: |
    Session store and orchestration front-end for the AutWit manual-QA copilot.

    Design invariants:
      - This service is the ONLY writer to the autwit Postgres schema.
      - Every operation that calls the orchestrator or takes >1s is a "run".
      - Runs are submit-only: POST returns 202 with a run_id. Never block.
      - Truth is always GET /sessions/{id}. SSE is a hint, not a guarantee.
      - The orchestrator returns everything inline (no pre-signed URIs yet).
servers:
  - url: http://localhost:8080/api/v1
    description: Local
  - url: https://autwit-copilot.internal/api/v1
    description: Internal
...
paths:
...
  /sessions:
    post:
      tags: [sessions]
      summary: Start a test session
      operationId: createSession
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSessionRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/BadRequest' }

    get:
      tags: [sessions]
      summary: List sessions
      operationId: listSessions
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/SessionStatus' }
        - name: tester_id
          in: query
          schema: { type: string }
        - name: env
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionPage' }
  /sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [sessions, timeline]
      summary: Get session with full timeline
      description: |
        THE source of truth. The UI hydrates from this on load and refetches
        after any run.succeeded / run.failed notification.
      operationId: getSession
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [sessions]
      summary: Update session metadata (title, tags, subjects, retention)
      operationId: updateSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSessionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
...
  /planning/sessions:
    get:
      tags: [planning]
      summary: A tester's recent planning sessions (the resume list)
      operationId: listPlanningSessions
      parameters:
        - name: tester_id
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlanningSessionView' }
    post:
      tags: [planning]
      summary: Create a planning session (+ its first project)
      description: |
        A resumable, single-tester context whose OpenAI lineage the next generation reuses.
        Creates the session and its first project in one call.
      operationId: createPlanningSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanningSessionRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatePlanningSessionResponse' }
  /planning/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [planning]
      summary: Resume a session — its projects + history timeline
      operationId: getPlanningSession
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanningSessionDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects:
    get:
      tags: [planning]
      summary: List planning projects
      operationId: listPlanningProjects
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlanningProject' }
    post:
      tags: [planning]
      summary: Create a planning project (Step 1 — feature reference)
      operationId: createPlanningProject
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanningProjectRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanningProject' }
  /planning/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [planning]
      summary: Get a planning project
      operationId: getPlanningProject
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanningProject' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/documents:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [planning]
      summary: List a project's source documents
      operationId: listPlanningDocuments
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SourceDocumentView' }
    post:
      tags: [planning]
      summary: Add a pasted (or client-supplied text) document (Step 1)
      description: |
        JSON path for pasted text (source_type=paste). File uploads — including PDF/DOCX/XLSX,
        which are parsed server-side — go to POST .../documents/upload as multipart.
      operationId: addPlanningDocument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddDocumentRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceDocumentView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/documents/upload:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [planning]
      summary: Upload a document file, parsed server-side (Step 1)
      description: |
        Multipart file upload. The bytes are parsed by Tika — PDF (PDFBox), DOCX/XLSX (POI),
        and text/markdown/csv/html — into the corpus text. A scanned/image-only PDF or an
        empty/encrypted file is a 400 (we do not OCR).
      operationId: uploadPlanningDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceDocumentView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/BadRequest' }
  /planning/projects/{projectId}/documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: documentId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [planning]
      summary: Select or deselect a document for generation
      operationId: selectPlanningDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selected]
              properties:
                selected: { type: boolean }
      responses:
        '204': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [planning]
      summary: Remove a document from the corpus
      operationId: deletePlanningDocument
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/jira-search:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: query
        in: query
        schema: { type: string }
      - name: project
        in: query
        schema: { type: string }
      - name: max
        in: query
        schema: { type: integer }
    get:
      tags: [planning]
      summary: Search Jira over MCP for candidate issues (Step 2)
      operationId: jiraSearch
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlanningCandidate' }
  /planning/projects/{projectId}/confluence-search:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: query
        in: query
        schema: { type: string }
      - name: space
        in: query
        schema: { type: string }
      - name: max
        in: query
        schema: { type: integer }
    get:
      tags: [planning]
      summary: Search Confluence over MCP for candidate pages (Step 2)
      operationId: confluenceSearch
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlanningCandidate' }
  /planning/projects/{projectId}/fetch:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [planning]
      summary: Fetch selected Jira/Confluence context over MCP (Step 2)
      description: |
        Pulls the selected items' full bodies, persists each as a source document,
        and returns a timestamped console log for the fetch activity view.
      operationId: fetchPlanningContext
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FetchRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FetchResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/test-plan:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [planning]
      summary: The latest generated test plan (Step 3)
      operationId: getLatestTestPlan
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestPlanView' }
        '204': { description: No plan generated yet }
    post:
      tags: [planning]
      summary: Generate a test plan from the selected corpus (Step 3)
      description: Async — returns 202 with a generation id to poll.
      operationId: generateTestPlan
      responses:
        '202':
          description: Enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/test-data:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [planning]
      summary: Generate per-scenario test data (Step 4)
      description: Async — returns 202 with a generation id to poll.
      operationId: generateTestData
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateDataRequest' }
      responses:
        '202':
          description: Enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationView' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/generations:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [planning]
      summary: List a project's generations
      operationId: listGenerations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GenerationView' }
  /planning/projects/{projectId}/generations/{generationId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/GenerationId'
    get:
      tags: [planning]
      summary: Get a generation's status (poll for completion)
      operationId: getGeneration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationView' }
        '404': { $ref: '#/components/responses/NotFound' }
  /planning/projects/{projectId}/generations/{generationId}/test-plan:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/GenerationId'
    get:
      tags: [planning]
      summary: The test plan produced by a generation
      operationId: getTestPlanByGeneration
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestPlanView' }
        '204': { description: Not ready or not a test-plan generation }
  /planning/projects/{projectId}/generations/{generationId}/test-data:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/GenerationId'
    get:
      tags: [planning]
      summary: The datasets produced by a generation (one per scenario)
      operationId: getTestData
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TestDatasetView' }
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Scoped to the session. Replaying a key returns the original run.
        Send on every POST. Double-clicking "Fulfil order" must not place
        two orders.
      schema: { type: string, maxLength: 200 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    SessionStatus:
      type: string
      enum: [active, ended, abandoned, archived]
    RetentionClass:
      type: string
      enum: [ephemeral, standard, pinned]
...
    CreateSessionRequest:
      type: object
      required: [tester_id, env]
      properties:
        tester_id: { type: string }
        env: { type: string, examples: [qa2, stage] }
        title: { type: string }
        tags:
          type: object
          additionalProperties: true
        subjects:
          $ref: '#/components/schemas/Subjects'
        retention_class:
          $ref: '#/components/schemas/RetentionClass'
        ttl_days:
          type: integer
          description: |
            Omit to use the server's configured default (autwit.session.default-ttl,
            currently 7d). Deliberately declares no `default:` here — the server owns
            that value, and a client that helpfully filled in a spec-declared 7 would
            silently override an operator who had configured 14.
    UpdateSessionRequest:
      type: object
      properties:
        title: { type: string }
        tags: { type: object, additionalProperties: true }
        subjects: { $ref: '#/components/schemas/Subjects' }
        retention_class: { $ref: '#/components/schemas/RetentionClass' }
...
    Session:
      type: object
      required: [session_id, correlation_id, tester_id, env, status, started_at]
      properties:
        session_id: { type: string, format: uuid }
        correlation_id:
          type: string
          description: Propagated downstream as X-Autwit-Correlation-Id
        tester_id: { type: string }
        env: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/SessionStatus' }
        retention_class: { $ref: '#/components/schemas/RetentionClass' }
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        tags: { type: object, additionalProperties: true }
        subjects: { $ref: '#/components/schemas/Subjects' }
    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/Session'
        - type: object
          required: [steps, milestones, snapshots, findings, active_runs]
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/Step' }
            milestones:
              type: array
              items: { $ref: '#/components/schemas/Milestone' }
            snapshots:
              type: array
              items: { $ref: '#/components/schemas/Snapshot' }
            comparisons:
              type: array
              items: { $ref: '#/components/schemas/Comparison' }
            findings:
              type: array
              items: { $ref: '#/components/schemas/Finding' }
            active_runs:
              type: array
              items: { $ref: '#/components/schemas/Run' }
            counts:
              type: object
              properties:
                artifacts: { type: integer }
                events: { type: integer }
                findings_by_severity:
                  type: object
                  additionalProperties: { type: integer }
    SessionPage:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items: { $ref: '#/components/schemas/Session' }
        next_cursor: { type: string }

    # ---- step
...
    PlanningSessionView:
      type: object
      required: [session_id, status, chainable]
      properties:
        session_id: { type: string, format: uuid }
        tester_id: { type: string, nullable: true }
        env: { type: string, nullable: true }
        title: { type: string, nullable: true }
        status: { type: string, enum: [active, archived] }
        chainable:
          type: boolean
          description: True once a generation has produced a lineage the next request reuses.
        created_at: { type: string, format: date-time }
        last_active_at: { type: string, format: date-time }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []